+++ modules/School_Setup/PortalPolls.php
@@ -348,36 +348,12 @@ function _makeOptionsInputs( $value, $name )
 
 	if ( $id == 'new' )
 	{
-		ob_start();
-		?>
-		<script>
-			var portalPollNewOption = function()
-			{
-				var table = document.getElementById('new-options-table'),
-					nbOptions = (table.rows.length - 1),
-					row = table.insertRow(nbOptions);
-
-				// Fill the cells.
-				function createCell(cell, tr, newId) {
-					cell.innerHTML = tr.cells[0].innerHTML;
-					reg = new RegExp('new' + (newId - 1),'g'); //g for global string
-					cell.innerHTML = cell.innerHTML.replace(reg, 'new' + newId);
-				}
-
-				// Insert table cells to the new row.
-				var tr = document.getElementById('new-option-' + nbOptions);
-
-				row.setAttribute('id', 'new-option-' + (nbOptions + 1));
-
-				createCell(row.insertCell(0), tr, nbOptions + 1);
-			};
-		</script>
-		<?php
-		$return .= ob_get_clean();
+		// @since 12.5 CSP remove unsafe-inline Javascript
+		$return .= '<script src="assets/js/csp/modules/schoolSetup/PortalPolls.js?v=12.5"></script>';
 
 		$return .= '<table class="widefat" id="new-options-table">' .
 			$value .
-			'<tr><td class="align-right"><a href="#" onclick="portalPollNewOption();return false;">' .
+			'<tr><td class="align-right"><a href="#!" class="onclick-add-new-question">' .
 			button( 'add' ) . ' ' . _( 'New Question' ) . '</a></tr></table>';
 	}
 	else